+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add present_value and future_value for a list of cash flows
 
+* Add net_present_value for an investment and a list of cash flows
+
 ## Changelog for Ex_Money v0.0.5 October 8, 2016
 
 ### Enhancements